// src/store/quizRoom/quizRoom.types.ts
export enum QUIZ_ROOM_TYPES {
  CURRENT_QUESTION = "CURRENT_QUESTION",
  EDIT_QUESTION_ARRAY_FOR_CHECKING_ANSWER = "QUESTION_ARRAY_FOR_CHECKING_ANSWER",
  SET_ORIGINAL_ARRAY = "SET_ORIGINAL_ARRAY",
  SET_QUIZ_TAKER_NAME = "SET_QUIZ_TAKER_NAME",
  SET_QUIZ_FINAL_SCORE = "SET_QUIZ_FINAL_SCORE",
  GET_QUESTIONS_FROM_DB_START = "GET_QUESTIONS_FROM_DB_START",
  GET_QUESTIONS_FROM_DB_SUCCESS = "GET_QUESTIONS_FROM_DB_SUCCESS",
  GET_QUESTIONS_FROM_DB_FAILED = "GET_QUESTIONS_FROM_DB_FAILED",
  UPLOAD_QUIZ_RESULT_TO_OWNER_DB_START = "UPLOAD_QUIZ_RESULT_TO_OWNER_DB_START",
  UPLOAD_QUIZ_RESULT_TO_OWNER_DB_SUCCESS = "UPLOAD_QUIZ_RESULT_TO_OWNER_DB_SUCCESS",
  UPLOAD_QUIZ_RESULT_TO_OWNER_DB_FAILED = "UPLOAD_QUIZ_RESULT_TO_OWNER_DB_FAILED",
  ADD_CURRENT_ARRAY_NUMBER = "ADD_CURRENT_ARRAY_NUMBER",
  REDUCE_CURRENT_ARRAY_NUMBER = "REDUCE_CURRENT_ARRAY_NUMBER",
  SET_CHECKED_VALUE = "SET_CHECKED_VALUE",
}